+++ R/create.R
@@ -108,6 +108,33 @@ create_project <- function(path,
 #'   [use_course()] for one-time download of all files in a Git repo, without
 #'   any local or remote Git operations.
 #'
+#' @section Using SSH Keys on Windows:
+#' If you are a Windows user who connects to GitHub using SSH, as opposed to
+#' HTTPS, you may need to explicitly specify the paths to your keys and register
+#' this credential in the current R session. This helps if git2r, which usethis
+#' uses for Git operations, does not automatically find your keys or handle your
+#' passphrase.
+#'
+#' In the snippet below, do whatever is necessary to make the paths correct,
+#' e.g., replace `<USERNAME>` with your Windows username. Omit the `passphrase`
+#' part if you don't have one. Replace `<OWNER/REPO>` with the appropriate
+#' GitHub specification. You get the idea.
+#'
+#' ```
+#' creds <- git2r::cred_ssh_key(
+#'   publickey  = "C:/Users/<USERNAME>/.ssh/id_rsa.pub",
+#'   privatekey = "C:/Users/<USERNAME>/.ssh/id_rsa",
+#'   passphrase = character(0)
+#' )
+#' use_git_protocol("ssh")
+#' use_git_credentials(credentials = creds)
+#'
+#' create_from_github(
+#'   repo_spec = "<OWNER/REPO>",
+#'   ...
+#' )
+#' ```
+#'
 #' @inheritParams create_package
 #' @param repo_spec GitHub repo specification in this form: `owner/repo`. The
 #'   `repo` part will be the name of the new local repo.