+++ tests/testthat/test-use-git.R
@@ -4,7 +4,7 @@ test_that('use_git_config(scope = "project) errors if project not using git', {
   # git2r::git2r::discover_repository() not working on R 3.1 (Travis)
   skip_if(getRversion() < 3.2)
   scoped_temporary_package()
-  expect_error(
+  expect_usethis_error(
     use_git_config(scope = "project", user.name = "USER.NAME"),
     "Cannot detect that project is already a Git repository"
   )
@@ -32,7 +32,7 @@ test_that("use_git_hook errors if project not using git", {
   # git2r::git2r::discover_repository() not working on R 3.1 (Travis)
   skip_if(getRversion() < 3.2)
   scoped_temporary_package()
-  expect_error(
+  expect_usethis_error(
     use_git_hook(
       "pre-commit",
       render_template("readme-rmd-pre-commit.sh")